requirements.txt
python-telegram-bot==13.7
spotipy
python-dotenv
yt-dlp
